requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.3
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.8
ordereddict==1.1
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
